% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TotalSpellTimeExceedingThreshold.R
\name{CST_TotalSpellTimeExceedingThreshold}
\alias{CST_TotalSpellTimeExceedingThreshold}
\title{Total Spell Time Exceeding Threshold}
\usage{
CST_TotalSpellTimeExceedingThreshold(
  data,
  threshold,
  spell,
  op = ">",
  start = NULL,
  end = NULL,
  time_dim = "ftime",
  ncores = NULL
)
}
\arguments{
\item{data}{An 's2dv_cube' object as provided by function \code{CST_Load} in 
package CSTools.}

\item{threshold}{An 's2dv_cube' object as output of a 'CST_' function in the 
same units as parameter 'data' and with the common dimensions of the element
'data' of the same length. A single scalar is also possible. If 
\code{timd_dim} is in the dimension (with the same length as \code{data}), 
the comparison will be done day by day.}

\item{spell}{A scalar indicating the minimum length of the spell.}

\item{op}{An operator '>' (by default), '<', '>=' or '<='.}

\item{start}{An optional parameter to defined the initial date of the period 
to select from the data by providing a list of two elements: the initial 
date of the period and the initial month of the period. By default it is set
to NULL and the indicator is computed using all the data provided in 
\code{data}.}

\item{end}{An optional parameter to defined the final date of the period to 
select from the data by providing a list of two elements: the final day of 
the period and the final month of the period. By default it is set to NULL 
and the indicator is computed using all the data provided in \code{data}.}

\item{time_dim}{A character string indicating the name of the dimension to 
compute the indicator. By default, it is set to 'ftime'. More than one 
dimension name matching the dimensions provided in the object 
\code{data$data} can be specified.}

\item{ncores}{An integer indicating the number of cores to use in parallel 
computation.}
}
\value{
An 's2dv_cube' object containing the indicator in the element 
\code{data}.
}
\description{
The number of days (when daily data is provided) that are part of a spell 
(defined by its minimum length e.g. 6 consecutive days) that exceed (or not 
exceed) a threshold are calculated with \code{TotalSpellTimeExceedingThreshold}.
This function allows to compute indicators widely used in Climate Services, 
such as:
\itemize{
\code{WSDI}{Warm Spell Duration Index that count the total number of days 
            with at least 6 consecutive days when the daily temperature 
            maximum exceeds its 90th percentile.}
}
This function requires the data and the threshold to be in the same units. The 
90th percentile can be translate into absolute values given a reference dataset 
using function \code{Threshold} or the data can be transform into probabilites 
by using function \code{AbsToProbs}. See section @examples.
}
\examples{
exp <- NULL
exp$data <- array(rnorm(5 * 3 * 214 * 2)*23,
                 c(member = 5, sdate = 3, ftime = 214, lon = 2)) 
exp$Dates$start <- c(seq(as.Date("01-05-2000", format = "\%d-\%m-\%Y"), 
                      as.Date("30-11-2000", format = "\%d-\%m-\%Y"), by = 'day'),
                  seq(as.Date("01-05-2001", format = "\%d-\%m-\%Y"), 
                      as.Date("30-11-2001", format = "\%d-\%m-\%Y"), by = 'day'),
                  seq(as.Date("01-05-2002", format = "\%d-\%m-\%Y"), 
                      as.Date("30-11-2002", format = "\%d-\%m-\%Y"), by = 'day'))
class(exp) <- 's2dv_cube'
TTSET <- CST_TotalSpellTimeExceedingThreshold(exp, threshold = 23, spell = 3)

}
\seealso{
[Threshold()] and [AbsToProbs()].
}
