% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelectPeriodOnDates.R
\name{SelectPeriodOnDates}
\alias{SelectPeriodOnDates}
\title{Select a period on Dates}
\usage{
SelectPeriodOnDates(dates, start, end, time_dim = "time", ncores = NULL)
}
\arguments{
\item{dates}{An array of dates with named dimensions.}

\item{start}{An optional parameter to defined the initial date of the period 
to select from the data by providing a list of two elements: the initial 
date of the period and the initial month of the period.}

\item{end}{An optional parameter to defined the final date of the period to 
select from the data by providing a list of two elements: the final day of 
the period and the final month of the period.}

\item{time_dim}{A character string indicating the name of the dimension to 
compute select the dates. By default, it is set to 'time'. More than one 
dimension name matching the dimensions provided in the object 
\code{data$data} can be specified.}

\item{ncores}{An integer indicating the number of cores to use in parallel 
computation.}
}
\value{
A multidimensional array with named dimensions containing the subset of 
the vector dates during the period requested from \code{start} to \code{end}.
}
\description{
Auxiliary function to subset dates for a specific period.
}
\examples{
Dates <- c(seq(as.Date("01-05-2000", format = "\%d-\%m-\%Y"), 
              as.Date("30-11-2000", format = "\%d-\%m-\%Y"), by = 'day'),
          seq(as.Date("01-05-2001", format = "\%d-\%m-\%Y"), 
              as.Date("30-11-2001", format = "\%d-\%m-\%Y"), by = 'day'),
          seq(as.Date("01-05-2002", format = "\%d-\%m-\%Y"), 
              as.Date("30-11-2002", format = "\%d-\%m-\%Y"), by = 'day'))
dim(Dates) <- c(time = 214, sdate = 3)
Period <- SelectPeriodOnDates(Dates, start = list(21, 6), end = list(21, 9))
}
